% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econid-package.R
\docType{package}
\name{econid-package}
\alias{econid}
\alias{econid-package}
\title{econid: Economic Entity Identifier Standardization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides utility functions for standardizing economic entity (economy, aggregate, institution, etc.) name and id in economic datasets such as those published by the International Monetary Fund and World Bank. Aims to facilitate consistent data analysis, reporting, and joining across datasets. Used as a foundational building block in the 'econdataverse' family of packages (\url{https://www.econdataverse.org}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://teal-insights.github.io/r-econid/}
  \item \url{https://github.com/Teal-Insights/r-econid}
  \item Report bugs at \url{https://github.com/Teal-Insights/r-econid/issues}
}

}
\author{
\strong{Maintainer}: L. Teal Emery \email{lte@tealinsights.com}

Authors:
\itemize{
  \item Christopher C. Smith \email{christopher.smith@promptlytechnologies.com}
}

Other contributors:
\itemize{
  \item Christoph Scheuch \email{christoph@tidy-intelligence.com} [contributor]
  \item Teal Insights \email{lte@tealinsights.com} [copyright holder]
}

}
\keyword{internal}
