% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_patterns.R
\name{reset_custom_entity_patterns}
\alias{reset_custom_entity_patterns}
\title{Reset custom entity patterns}
\usage{
reset_custom_entity_patterns()
}
\value{
Invisibly returns NULL.
}
\description{
This function resets all custom entity patterns that have been added
during the current R session.
}
\examples{
add_entity_pattern("EU", "European Union", "economy")
reset_custom_entity_patterns()
patterns <- list_entity_patterns()
print(patterns[patterns$entity_id == "EU", ])

}
